requirements.txt
selenium
pytest
pyyaml
telepot
pillow